requirements.txt
Django==1.8.6
dj-database-url==0.3.0
django-tinymce==2.0.5
django-tagging==0.4.0
Pillow==3.0.0
wheel==0.24.0
